Reorganize some types
