+++ crates/ibc/src/core/ics02_client/events.rs
@@ -15,8 +15,6 @@ const CREATE_CLIENT_EVENT: &str = "create_client";
 const UPDATE_CLIENT_EVENT: &str = "update_client";
 const CLIENT_MISBEHAVIOUR_EVENT: &str = "client_misbehaviour";
 const UPGRADE_CLIENT_EVENT: &str = "upgrade_client";
-const UPGRADE_CHAIN_EVENT: &str = "upgrade_chain";
-const UPGRADE_CLIENT_PROPOSAL: &str = "upgrade_client_proposal";
 
 /// The content of the `key` field for the attribute containing the client identifier.
 pub const CLIENT_ID_ATTRIBUTE_KEY: &str = "client_id";
@@ -32,15 +30,6 @@ pub const CONSENSUS_HEIGHTS_ATTRIBUTE_KEY: &str = "consensus_heights";
 
 /// The content of the `key` field for the header in update client event.
 pub const HEADER_ATTRIBUTE_KEY: &str = "header";
-
-/// The content of the `key` field for the attribute containing the upgrade plan name.
-const KEY_UPGRADE_STORE_ATTRIBUTE_KEY: &str = "upgrade_store";
-
-/// The content of the `key` field for the attribute containing the upgrade plan height.
-const UPGRADE_PLAN_HEIGHT_ATTRIBUTE_KEY: &str = "upgrade_plan_height";
-
-/// The content of the `key` field for the attribute containing the upgrade plan title.
-const UPGRADE_PLAN_TITLE_ATTRIBUTE_KEY: &str = "title";
 
 #[cfg_attr(
     feature = "parity-scale-codec",
@@ -168,82 +157,6 @@ impl From<HeaderAttribute> for abci::EventAttribute {
             String::from_utf8(hex::encode(attr.header.value)).unwrap(),
         )
             .into()
-    }
-}
-
-#[cfg_attr(
-    feature = "parity-scale-codec",
-    derive(
-        parity_scale_codec::Encode,
-        parity_scale_codec::Decode,
-        scale_info::TypeInfo
-    )
-)]
-#[cfg_attr(
-    feature = "borsh",
-    derive(borsh::BorshSerialize, borsh::BorshDeserialize)
-)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug, From, PartialEq, Eq)]
-struct UpgradeStoreAttribute {
-    upgrade_store: String,
-}
-
-impl From<UpgradeStoreAttribute> for abci::EventAttribute {
-    fn from(attr: UpgradeStoreAttribute) -> Self {
-        (KEY_UPGRADE_STORE_ATTRIBUTE_KEY, attr.upgrade_store).into()
-    }
-}
-
-#[cfg_attr(
-    feature = "parity-scale-codec",
-    derive(
-        parity_scale_codec::Encode,
-        parity_scale_codec::Decode,
-        scale_info::TypeInfo
-    )
-)]
-#[cfg_attr(
-    feature = "borsh",
-    derive(borsh::BorshSerialize, borsh::BorshDeserialize)
-)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug, From, PartialEq, Eq)]
-struct UpgradePlanHeightAttribute {
-    plan_height: u64,
-}
-
-impl From<UpgradePlanHeightAttribute> for abci::EventAttribute {
-    fn from(attr: UpgradePlanHeightAttribute) -> Self {
-        (
-            UPGRADE_PLAN_HEIGHT_ATTRIBUTE_KEY,
-            attr.plan_height.to_string(),
-        )
-            .into()
-    }
-}
-
-#[cfg_attr(
-    feature = "parity-scale-codec",
-    derive(
-        parity_scale_codec::Encode,
-        parity_scale_codec::Decode,
-        scale_info::TypeInfo
-    )
-)]
-#[cfg_attr(
-    feature = "borsh",
-    derive(borsh::BorshSerialize, borsh::BorshDeserialize)
-)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug, From, PartialEq, Eq)]
-struct UpgradePlanTitleAttribute {
-    title: String,
-}
-
-impl From<UpgradePlanTitleAttribute> for abci::EventAttribute {
-    fn from(attr: UpgradePlanTitleAttribute) -> Self {
-        (UPGRADE_PLAN_TITLE_ATTRIBUTE_KEY, attr.title).into()
     }
 }
 
@@ -496,86 +409,6 @@ impl From<UpgradeClient> for abci::Event {
                 u.client_type.into(),
                 u.consensus_height.into(),
             ],
-        }
-    }
-}
-
-#[cfg_attr(
-    feature = "parity-scale-codec",
-    derive(
-        parity_scale_codec::Encode,
-        parity_scale_codec::Decode,
-        scale_info::TypeInfo
-    )
-)]
-#[cfg_attr(
-    feature = "borsh",
-    derive(borsh::BorshSerialize, borsh::BorshDeserialize)
-)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct UpgradeChain {
-    plan_height: UpgradePlanHeightAttribute,
-    upgrade_store: UpgradeStoreAttribute,
-}
-
-impl UpgradeChain {
-    pub fn new(plan_height: u64, upgrade_store: String) -> Self {
-        Self {
-            plan_height: UpgradePlanHeightAttribute::from(plan_height),
-            upgrade_store: UpgradeStoreAttribute::from(upgrade_store),
-        }
-    }
-    pub fn event_type(&self) -> &str {
-        UPGRADE_CHAIN_EVENT
-    }
-}
-
-impl From<UpgradeChain> for abci::Event {
-    fn from(u: UpgradeChain) -> Self {
-        Self {
-            kind: UPGRADE_CHAIN_EVENT.to_owned(),
-            attributes: vec![u.plan_height.into(), u.upgrade_store.into()],
-        }
-    }
-}
-
-#[cfg_attr(
-    feature = "parity-scale-codec",
-    derive(
-        parity_scale_codec::Encode,
-        parity_scale_codec::Decode,
-        scale_info::TypeInfo
-    )
-)]
-#[cfg_attr(
-    feature = "borsh",
-    derive(borsh::BorshSerialize, borsh::BorshDeserialize)
-)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct UpgradeClientProposal {
-    plan_title: UpgradePlanTitleAttribute,
-    plan_height: UpgradePlanHeightAttribute,
-}
-
-impl UpgradeClientProposal {
-    pub fn new(plan_title: String, plan_height: u64) -> Self {
-        Self {
-            plan_title: UpgradePlanTitleAttribute::from(plan_title),
-            plan_height: UpgradePlanHeightAttribute::from(plan_height),
-        }
-    }
-    pub fn event_type(&self) -> &str {
-        UPGRADE_CLIENT_PROPOSAL
-    }
-}
-
-impl From<UpgradeClientProposal> for abci::Event {
-    fn from(u: UpgradeClientProposal) -> Self {
-        Self {
-            kind: UPGRADE_CLIENT_PROPOSAL.to_owned(),
-            attributes: vec![u.plan_title.into(), u.plan_height.into()],
         }
     }
 }
