+++ crates/ibc/src/hosts/tendermint/upgrade_proposal/mod.rs
@@ -1,5 +1,9 @@
 mod context;
+mod events;
+mod plan;
 mod proposal;
 
 pub use context::{UpgradeExecutionContext, UpgradeValidationContext};
-pub use proposal::{Plan, UpgradeProposal};
+pub use events::{UpgradeChain, UpgradeClientProposal};
+pub use plan::Plan;
+pub use proposal::UpgradeProposal;
