+++ crates/ibc/src/hosts/tendermint/upgrade_proposal/proposal.rs
@@ -1,11 +1,11 @@
-//! Definition of domain `UpgradeProposal` and `Plan` types for handling upgrade client proposals
+//! Definition of domain `UpgradeProposal` type for handling upgrade client proposal
 
 use alloc::string::{String, ToString};
-use ibc_proto::cosmos::upgrade::v1beta1::Plan as RawPlan;
 use ibc_proto::google::protobuf::Any;
 use ibc_proto::ibc::core::client::v1::UpgradeProposal as RawUpgradeProposal;
 use ibc_proto::protobuf::Protobuf;
 
+use super::Plan;
 use crate::core::ics02_client::error::UpgradeClientError;
 
 #[derive(Clone, Debug)]
@@ -66,64 +66,6 @@ impl From<UpgradeProposal> for RawUpgradeProposal {
             description: value.description,
             plan: Some(value.plan.into()),
             upgraded_client_state: Some(value.upgraded_client_state),
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct Plan {
-    pub name: String,
-    pub height: u64,
-    pub info: String,
-}
-
-impl Protobuf<RawPlan> for Plan {}
-
-impl TryFrom<RawPlan> for Plan {
-    type Error = UpgradeClientError;
-
-    fn try_from(raw: RawPlan) -> Result<Self, Self::Error> {
-        if raw.name.is_empty() {
-            return Err(UpgradeClientError::InvalidUpgradePlan {
-                reason: "name field cannot be empty".to_string(),
-            });
-        }
-
-        #[allow(deprecated)]
-        if raw.time.is_some() {
-            return Err(UpgradeClientError::InvalidUpgradePlan {
-                reason: "time field must be empty".to_string(),
-            });
-        }
-
-        #[allow(deprecated)]
-        if raw.upgraded_client_state.is_some() {
-            return Err(UpgradeClientError::InvalidUpgradePlan {
-                reason: "upgraded_client_state field must be empty".to_string(),
-            });
-        }
-
-        Ok(Self {
-            name: raw.name,
-            height: u64::try_from(raw.height).map_err(|_| {
-                UpgradeClientError::InvalidUpgradePlan {
-                    reason: "height plan overflow".to_string(),
-                }
-            })?,
-            info: raw.info,
-        })
-    }
-}
-
-impl From<Plan> for RawPlan {
-    fn from(value: Plan) -> Self {
-        #[allow(deprecated)]
-        Self {
-            name: value.name,
-            height: i64::try_from(value.height).expect("height overflow"),
-            info: value.info,
-            time: None,
-            upgraded_client_state: None,
         }
     }
 }
